+++ src/main/java/com/foodtimetest/message/controller/MessageServlet.java
@@ -1,0 +1,5 @@
+package com.foodtimetest.message.controller;
+
+public class MessageServlet {
+
+}
