+++ src/main/java/com/foodtimetest/message/model/MessageDAO_interface.java
@@ -1,4 +1,4 @@
-package com.foodtimetest.message;
+package com.foodtimetest.message.model;
 
 import java.util.*;
